remove from cart now relies on user id

+++ db/delete_from_cart.sql
@@ -1,2 +1,3 @@
 delete from cart 
 where product_id = ${id}
+and user_id = ${userId}
